LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/container-queries/anchored-fallback-display-change-expected.txt


PASS @container anchored() without applied fallback
FAIL @container anchored() with fallback applied assert_equals: Fallback applies display:block expected 100 but got 0
